working reusable component
